Add Qwen QwQ
