+++ lmdeploy/serve/openai/reasoning_parser/__init__.py
@@ -1,4 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+from .deepseek_r1_reasoning_parser import DeepSeekR1ReasoningParser
+from .qwen_qwq_reasoning_parser import QwenQwQReasoningParser
 from .reasoning_parser import ReasoningParser, ReasoningParserManager
 
-__all__ = ['ReasoningParser', 'ReasoningParserManager']
+__all__ = ['ReasoningParser', 'ReasoningParserManager', 'DeepSeekR1ReasoningParser', 'QwenQwQReasoningParser']
